+++ samples/golf/src/golf/CollisionMesh.hpp
@@ -87,7 +87,9 @@ private:
     std::vector<std::unique_ptr<btTriangleIndexVertexArray>> m_groundVertices;
 
     std::vector<float> m_vertexData;
-    std::vector<std::vector<std::uint32_t>> m_indexData;
+    //hmm I wish we could query the index data for its type
+    //and automatically select the correct one here...
+    std::vector<std::vector<std::uint16_t>> m_indexData;
 
 
     void initCollisionWorld();
